backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
openai>=1.12.0
firebase-admin==6.2.0
pyjwt==2.8.0
# NUEVAS DEPENDENCIAS
flask-limiter==3.5.0
marshmallow==3.20.1
flask-talisman==1.1.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0